+++ src/main/java/com/dnd/gongmuin/answer/dto/AnswerDetailResponse.java
@@ -1,6 +1,6 @@
 package com.dnd.gongmuin.answer.dto;
 
-import com.dnd.gongmuin.question_post.dto.MemberInfo;
+import com.dnd.gongmuin.question_post.dto.response.MemberInfo;
 
 public record AnswerDetailResponse(
 	Long answerId,
