+++ src/main/java/com/dnd/gongmuin/question_post/dto/QuestionPostMapper.java
@@ -6,6 +6,9 @@ import com.dnd.gongmuin.member.domain.JobGroup;
 import com.dnd.gongmuin.member.domain.Member;
 import com.dnd.gongmuin.question_post.domain.QuestionPost;
 import com.dnd.gongmuin.question_post.domain.QuestionPostImage;
+import com.dnd.gongmuin.question_post.dto.request.RegisterQuestionPostRequest;
+import com.dnd.gongmuin.question_post.dto.response.MemberInfo;
+import com.dnd.gongmuin.question_post.dto.response.QuestionPostDetailResponse;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
