+++ src/main/java/com/dnd/gongmuin/question_post/dto/response/MemberInfo.java
@@ -1,4 +1,4 @@
-package com.dnd.gongmuin.question_post.dto;
+package com.dnd.gongmuin.question_post.dto.response;
 
 public record MemberInfo(
 	Long memberId,
