+++ src/main/java/com/dnd/gongmuin/question_post/dto/response/QuestionPostDetailResponse.java
@@ -1,4 +1,4 @@
-package com.dnd.gongmuin.question_post.dto;
+package com.dnd.gongmuin.question_post.dto.response;
 
 import java.util.List;
