+++ src/test/java/com/dnd/gongmuin/question_post/service/QuestionPostServiceTest.java
@@ -19,8 +19,8 @@ import com.dnd.gongmuin.common.fixture.MemberFixture;
 import com.dnd.gongmuin.common.fixture.QuestionPostFixture;
 import com.dnd.gongmuin.member.domain.Member;
 import com.dnd.gongmuin.question_post.domain.QuestionPost;
-import com.dnd.gongmuin.question_post.dto.QuestionPostDetailResponse;
-import com.dnd.gongmuin.question_post.dto.RegisterQuestionPostRequest;
+import com.dnd.gongmuin.question_post.dto.response.QuestionPostDetailResponse;
+import com.dnd.gongmuin.question_post.dto.request.RegisterQuestionPostRequest;
 import com.dnd.gongmuin.question_post.repository.QuestionPostRepository;
 
 @DisplayName("[QuestionPostService 테스트]")
